ANZ - Removed unnecessary sponsor categories
The extra sponsor list categories copied over from SUGCON EU have been removed from the presentation details of the SUGCON ANZ Sponsors page. I have kept the lists in the data folder for future use in other years.

While in there, I also removed the 'Blastic' reference in the Video Recording sponsors list, as that was a copy from EU and not accurate for ANZ.